Update Patch request so that it increases the helpful vote count for a single listing image by 1 when called
