+++ server/database/index.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
 const mongoUri = 'mongodb://localhost/Listing';
 
-const db = mongoose.connect(mongoUri);
+const db = mongoose.connect(mongoUri, { useNewUrlParser: true });
 
 module.exports = db;
